minor adjustment to table

+++ src/main/scala/org/brzy/webapp/controller/Table.scala
@@ -5,7 +5,7 @@ package org.brzy.webapp.controller
  *
  * @author Michael Fortin
  */
-case class Table[T](list: List[T], pageStart: Int, pageLimit: Int, total: Int) {
+case class Table[T](list: List[T], pageStart: Int, pageLimit: Int, total: Long) {
 
   def hasPrevious: Boolean = pageStart - pageLimit >= 0
 
@@ -18,7 +18,7 @@ case class Table[T](list: List[T], pageStart: Int, pageLimit: Int, total: Int) {
 
   def currentPage: Int = pageStart
 
-  def pagesAfter: Array[Int] = {
+  def pagesAfter: Array[Long] = {
     val x = for (i <- pageStart+pageLimit to total; if i % pageLimit == 0) yield i
     x.toArray
   }
